capture existing Products/Restaurants as Test*, point tests at Test*
