--- artifacts/Products.1/Products.recipes
@@ -1,20 +1,0 @@
-// @license
-// Copyright (c) 2017 Google Inc. All rights reserved.
-// This code may only be used under the BSD style license found at
-// http://polymer.github.io/LICENSE.txt
-// Code distributed by Google as part of this project is also
-// subject to an additional IP rights grant found at
-// http://polymer.github.io/PATENTS.txt
-
-// here for demo-flow-test.js only
-import '../Common/List.recipes'
-
-import 'ShowProducts.recipes'
-import 'Recommend.recipes'
-import 'Manufacturer.recipes'
-import 'Gifts.recipes'
-import 'Interests.recipes'
-
-import '../Demo/Browse.manifest'
-import '../Demo/ClairesWishlist.manifest'
-
