--- artifacts/Restaurants.1/Restaurants.recipes
@@ -1,58 +1,0 @@
-// Copyright (c) 2017 Google Inc. All rights reserved.
-// This code may only be used under the BSD style license found at
-// http://polymer.github.io/LICENSE.txt
-// Code distributed by Google as part of this project is also
-// subject to an additional IP rights grant found at
-// http://polymer.github.io/PATENTS.txt
-
-import 'Restaurant.schema'
-
-//
-
-import '../Places/Places.recipes'
-import '../Places/GeoCoordinates.schema'
-
-particle FindRestaurants in 'source/FindRestaurants.js'
-  in GeoCoordinates location
-  inout [Restaurant] restaurants
-
-recipe FindRestaurants
-  create #tiles as restaurants
-  FindRestaurants
-    restaurants = restaurants
-
-//
-
-particle RestaurantTile in 'source/RestaurantTile.js'
-  in Restaurant restaurant
-  consume tile
-    provide annotation
-
-recipe ShowRestaurantTile
-  RestaurantTile
-
-//
-
-particle RestaurantDetail in 'source/RestaurantDetail.js'
-  in Restaurant selected
-  consume content
-    //provide action
-
-recipe RestaurantDetail
-  RestaurantDetail
-
-//
-
-import '../Events/Event.schema'
-
-particle ReservationAnnotation in 'source/ReservationAnnotation.js'
-  in [Restaurant] list
-  inout Event event
-  consume annotation
-
-recipe RestaurantReservation
-  create #event as event
-  use as restaurants
-  ReservationAnnotation
-    event = event
-    list = restaurants
