+++ artifacts/Restaurants/source/RestaurantTile.js
