+++ artifacts/TestProducts/Products.recipes
@@ -1,14 +1,17 @@
 // @license
-// Copyright (c) 2018 Google Inc. All rights reserved.
+// Copyright (c) 2017 Google Inc. All rights reserved.
 // This code may only be used under the BSD style license found at
 // http://polymer.github.io/LICENSE.txt
 // Code distributed by Google as part of this project is also
 // subject to an additional IP rights grant found at
 // http://polymer.github.io/PATENTS.txt
 
-import 'Product.schema'
+import 'ShowProducts.recipes'
+import 'Recommend.recipes'
+import 'Manufacturer.recipes'
+import 'Gifts.recipes'
+import 'Interests.recipes'
 
-particle ShowProduct in 'source/ShowProduct.js'
-  in Product product
-  consume item
+import '../Demo/Browse.manifest'
+import '../Demo/ClairesWishlist.manifest'
