+++ artifacts/TestProducts/source/ManufacturerInfo.js
