+++ artifacts/TestProducts/test/ProductIsBook.manifest
