+++ artifacts/TestRestaurants/FavoriteFoodAnnotation.manifest
