+++ artifacts/TestRestaurants/ReservationAnnotation.manifest
