+++ artifacts/TestRestaurants/Restaurant.schema
