+++ artifacts/TestRestaurants/RestaurantDetail.manifest
