+++ artifacts/TestRestaurants/RestaurantMasterDetail.manifest
