+++ artifacts/TestRestaurants/Restaurants.recipes
@@ -1,0 +1,60 @@
+// Copyright (c) 2017 Google Inc. All rights reserved.
+// This code may only be used under the BSD style license found at
+// http://polymer.github.io/LICENSE.txt
+// Code distributed by Google as part of this project is also
+// subject to an additional IP rights grant found at
+// http://polymer.github.io/PATENTS.txt
+
+import 'FindRestaurants.manifest'
+import 'RestaurantMasterDetail.manifest'
+import 'RestaurantList.manifest'
+import 'RestaurantDetail.manifest'
+import '../Places/ExtractLocation.manifest'
+
+recipe &displayRestaurants
+  ? as list
+  RestaurantMasterDetail.selected -> RestaurantList.selected
+  RestaurantDetail.selected -> RestaurantList.selected
+  RestaurantMasterDetail.list -> RestaurantList.list
+  RestaurantList
+    list <- list
+
+recipe
+  create as location
+  ExtractLocation
+    location -> location
+
+recipe &nearbyRestaurants
+  create #restaurants #nosync as restaurants
+  FindRestaurants
+    restaurants = restaurants
+
+import '../Events/Event.schema'
+import '../Events/PartySize.manifest'
+import 'ReservationForm.manifest'
+import 'ReservationAnnotation.manifest'
+
+recipe &makeReservation
+  create #event as event
+  use #restaurants as list
+  ReservationForm
+    event = event
+  ReservationAnnotation
+    event = event
+    list = list
+  PartySize
+    event = event
+
+import '../Events/Calendar.manifest'
+
+recipe
+  Calendar
+
+import 'FavoriteFoodAnnotation.manifest'
+
+recipe
+  map #favorite as food
+  use as restaurants
+  FavoriteFoodAnnotation
+    restaurants <- restaurants
+    food <- food
