+++ artifacts/TestRestaurants/source/FavoriteFoodAnnotation.js
