+++ artifacts/canonical.manifest
@@ -18,8 +18,8 @@ import 'Common/Common.recipes'
 //import 'People/People.recipes'
 import 'Places/Places.recipes'
 //import 'Profile/Profile.recipes'
-//import 'Products.1/Products.recipes'
-import 'Restaurants.1/Restaurants.recipes'
+//import 'Products/Products.recipes'
+import 'Restaurants/Restaurants.recipes'
 //import 'Social/Social.recipes'
 //import 'VideoPlayer/VideoPlayer.recipes'
 //import 'Words/Words.recipes'
