+++ runtime/test/particles/artifacts/products-test.recipes
@@ -7,9 +7,9 @@
 // http://polymer.github.io/PATENTS.txt
 
 import '../../../../artifacts/Common/List.manifest'
-import '../../../../artifacts/Products/test/ProductFilter.manifest'
-import '../../../../artifacts/Products/test/ProductIsBook.manifest'
-import '../../../../artifacts/Products/ShowProducts.recipes'
+import '../../../../artifacts/TestProducts/test/ProductFilter.manifest'
+import '../../../../artifacts/TestProducts/test/ProductIsBook.manifest'
+import '../../../../artifacts/TestProducts/ShowProducts.recipes'
 
 // Filter books and display them.
 recipe
